backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy==2.0.30
asyncpg==0.29.0
alembic==1.13.1
pydantic==2.7.1
pydantic-settings==2.3.1
pymupdf==1.24.2
python-docx==1.1.0
python-multipart==0.0.9
sentence-transformers==3.0.1
torch==2.3.1
playwright==1.44.0
twilio==9.1.0
apscheduler==3.10.4
httpx==0.27.0
aiofiles==23.2.1

# Fix #7: Rate limiting
slowapi==0.1.9

# Fix #20: Retry logic for scheduler
tenacity==8.3.0

# Testing
pytest==8.2.0
pytest-asyncio==0.23.6
anyio==4.4.0
